# Faster path drawing for the cairo backend (cairocffi only)