--- a/doc/api/backend_cairo_api.rst
+++ b/doc/api/backend_cairo_api.rst
@@ -2,7 +2,12 @@
 :mod:`matplotlib.backends.backend_cairo`
 ========================================
 
-.. automodule:: matplotlib.backends.backend_cairo
-   :members:
-   :undoc-members:
-   :show-inheritance:
+.. Building the docs requires either adding pycairo/cairocffi as docs build
+   dependency, or bumping the minimal numpy version to one that supports
+   MagicMocks (which does define `__index__`) as indices (recent numpys do, but
+   1.7.1 doesn't).
+
+.. .. automodule:: matplotlib.backends.backend_cairo
+..    :members:
+..    :undoc-members:
+..    :show-inheritance: