@QuLogic 2018-02-08
Generally, I don't see a problem with this change, but using your microbenchmark, I can't seem to reproduce the speedup you see:

```
$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post1307.dev0+g84008ea05
Backend: TkAgg
Average FPS: 19.965329

$ python3 examples/mplot3d/wire3d_animation.py
/usr/lib/python3.6/site-packages/cairocffi/surfaces.py:651: UserWarning: implicit cast from 'char *' to a different pointer type: will be forbidden in the future (check that the types are as you expect; use an explicit ffi.cast() if they are correct)
  ffi.cast('char*', address), format, width, height, stride)
Matplotlib version: 2.1.1.post1307.dev0+g84008ea05
Backend: GTK3Agg
Average FPS: 20.217351

$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post1307.dev0+g84008ea05
Backend: agg
Average FPS: 22.100502

$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post1307.dev0+g84008ea05
Backend: GTK3Cairo
Average FPS: 12.131168

$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post1307.dev0+g84008ea05
Backend: cairo
Average FPS: 70.399400
```
vs.
```
$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post18.dev0+gebcfd908c
Backend: TkAgg
Average FPS: 18.767531

$ python3 examples/mplot3d/wire3d_animation.py
/usr/lib/python3.6/site-packages/cairocffi/surfaces.py:651: UserWarning: implicit cast from 'char *' to a different pointer type: will be forbidden in the future (check that the types are as you expect; use an explicit ffi.cast() if they are correct)
  ffi.cast('char*', address), format, width, height, stride)
Matplotlib version: 2.1.1.post18.dev0+gebcfd908c
Backend: GTK3Agg
Average FPS: 19.497821

$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post18.dev0+gebcfd908c
Backend: agg
Average FPS: 21.796568

$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post18.dev0+gebcfd908c
Backend: GTK3Cairo
Average FPS: 12.018317

$ python3 examples/mplot3d/wire3d_animation.py
Matplotlib version: 2.1.1.post18.dev0+gebcfd908c
Backend: cairo
Average FPS: 68.990914
```