@anntzer 2018-02-08
Are you certain your gtk3cairo is using cairocffi rather than pycairo (which is not accelerated by this PR)?
I just tried and can still get the difference, ~8.7fps vs 12.2fps.