@QuLogic 2018-02-15
OK, I'm not really sure what happened, but I just tried again and I get 12.705427 vs. 17.470702 now.