@lazka 2018-05-28
Would it help you if pycairo had something like `cairo.Path.create_for_data(buffer, num_data)` ?