@lazka 2018-05-28
> I moved my efforts to just doing everything in C(++) (https://github.com/anntzer/mplcairo) but I guess such a method can't hurt in general...

ok

> Perhaps also worth (for you) to discuss this with cairocffi to present the same API in both projects.

yeah, but it's unmaintained atm